rename NHibernate orm feature to Default orm feature
