+++ src/blueprints/Do.Blueprints.Service.Application/Orm/Default/DefatultOrmExtensions.cs
@@ -1,0 +1,9 @@
+﻿using Do.Orm;
+using Do.Orm.Default;
+
+namespace Do;
+
+public static class DefatultOrmExtensions
+{
+    public static DefaultOrmFeature Default(this OrmConfigurator source) => new();
+}
