--- src/blueprints/Do.Blueprints.Service.Application/Orm/NHibernate/NHibernateOrmExtensions.cs
@@ -1,9 +1,0 @@
-﻿using Do.Orm;
-using Do.Orm.NHibernate;
-
-namespace Do;
-
-public static class NHibernateOrmExtensions
-{
-    public static NHibernateOrmFeature NHibernate(this OrmConfigurator source) => new();
-}
